fix for bug #1178
Apparently gcc 3.4 gives error when not casting a pointer to long.
Found and reported by Noselasd.

+++ source/blender/blenkernel/intern/exotic.c
@@ -1187,7 +1187,7 @@ static int iv_finddata(struct IvNode *iv, char *field, int fieldnr)
 							 * in a long string at Windows... so we copy 
 							 * the float to a new string then atof... */
 							 
-							i=strpbrk(cpa, ", \n")-cpa;
+							i= (long)(strpbrk(cpa, ", \n")-cpa);
 							
 							if (i>63) *fp= 0.0;
 							else {
